mosaicme/collector/requirements.txt
tweepy==3.3.0
pika==0.9.14
django-dotenv==1.3.0
celery==3.1.17
redis==2.10.3
requests==2.6.0
boto==2.37.0
supervisor==3.1.3
retrying==1.3.3
